+++ scripts/uniswap-v2-pairs-and-swaps.py
@@ -226,7 +226,7 @@ def main():
     json_rpc_url = os.environ["JSON_RPC_URL"]
     web3 = Web3(HTTPProvider(json_rpc_url, session=session))
 
-    # Enable faster ujson reads
+    # Enable faster orjson reads
     patch_web3(web3)
 
     web3.middleware_onion.clear()